refactor(episode): dedupe episode components and define superset
