--- projects/client/src/lib/components/media/card/MediaCover.svelte
@@ -1,12 +1,0 @@
-<script lang="ts">
-  import CardCover from "$lib/components/card/CardCover.svelte";
-  import type { CardCoverProps } from "$lib/components/card/CardCoverProps";
-
-  const { src, alt, tags: posterTags, isLoading }: CardCoverProps = $props();
-</script>
-
-<CardCover {src} {alt} {isLoading}>
-  {#snippet tags()}
-    {@render posterTags?.()}
-  {/snippet}
-</CardCover>
