+++ projects/client/src/lib/sections/lists/RecentlyWatchedList.svelte
@@ -29,6 +29,7 @@
         episode={media.episode}
         show={media.show}
         context="standalone"
+        type="default"
       />
     {/if}
