+++ projects/client/src/lib/sections/lists/UpcomingList.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import SectionList from "$lib/components/lists/section-list/SectionList.svelte";
   import * as m from "$lib/features/i18n/messages.ts";
+  import EpisodeCard from "./components/EpisodeCard.svelte";
   import FindShowsLink from "./components/FindShowsLink.svelte";
-  import UpcomingEpisodeCard from "./components/UpcomingEpisodeCard.svelte";
   import { useCalendarEpisodes } from "./stores/useCalendarEpisodes";
   import { mediaListHeightResolver } from "./utils/mediaListHeightResolver";
 
@@ -16,7 +16,7 @@
   --height-list={mediaListHeightResolver("episode")}
 >
   {#snippet item(entry)}
-    <UpcomingEpisodeCard episode={entry} show={entry.show} />
+    <EpisodeCard episode={entry} show={entry.show} type="upcoming" />
   {/snippet}
   {#snippet empty()}
     {#if !$isLoading}
