--- projects/client/src/lib/sections/lists/components/UpcomingEpisodeCard.svelte
@@ -1,70 +1,0 @@
-<script lang="ts">
-  import CardFooter from "$lib/components/card/CardFooter.svelte";
-  import EpisodeCard from "$lib/components/episode/card/EpisodeCard.svelte";
-  import EpisodeCover from "$lib/components/episode/card/EpisodeCover.svelte";
-  import { EpisodeIntlProvider } from "$lib/components/episode/EpisodeIntlProvider";
-  import EpisodeTimeTag from "$lib/components/episode/tags/EpisodeTimeTag.svelte";
-  import Link from "$lib/components/link/Link.svelte";
-  import Spoiler from "$lib/features/spoilers/components/Spoiler.svelte";
-  import { useEpisodeSpoilerImage } from "$lib/features/spoilers/useEpisodeSpoilerImage";
-  import type { EpisodeEntry } from "$lib/requests/models/EpisodeEntry";
-  import type { MediaEntry } from "$lib/requests/models/MediaEntry";
-  import { EPISODE_COVER_PLACEHOLDER } from "$lib/utils/constants";
-  import { UrlBuilder } from "$lib/utils/url/UrlBuilder";
-
-  type EpisodeProps = {
-    episode: EpisodeEntry;
-    show: MediaEntry;
-  };
-
-  const { show, episode }: EpisodeProps = $props();
-
-  const src = useEpisodeSpoilerImage({ episode, show });
-</script>
-
-<EpisodeCard>
-  <Link
-    focusable={false}
-    href={UrlBuilder.episode(show.slug, episode.season, episode.number)}
-  >
-    <EpisodeCover
-      i18n={EpisodeIntlProvider}
-      type={episode.type}
-      src={$src ?? EPISODE_COVER_PLACEHOLDER}
-      alt={`${show.title} - ${episode.title}`}
-    >
-      {#snippet tags()}
-        <EpisodeTimeTag>
-          {EpisodeIntlProvider.timestampText(episode.airDate)}
-        </EpisodeTimeTag>
-      {/snippet}
-    </EpisodeCover>
-  </Link>
-  <CardFooter>
-    <Link href={UrlBuilder.show(show.slug)}>
-      <p class="episode-show-title ellipsis">{show.title}</p>
-    </Link>
-    <p class="episode-title ellipsis small">
-      {episode.season}x{episode.number}
-      <Spoiler media={episode} {show} {episode} type="episode"
-        >- {episode.title}</Spoiler
-      >
-    </p>
-  </CardFooter>
-</EpisodeCard>
-
-<style>
-  .episode-show-title {
-    color: var(--color-text-primary);
-    margin: 0;
-
-    font-weight: 600;
-  }
-
-  .episode-title {
-    color: var(--color-text-secondary);
-    margin: 0;
-
-    font-weight: 500;
-  }
-</style>
